bot/requirements.txt
websockets>=12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0
python-telegram-bot>=20.7
pycryptodome>=3.19.0
